requirements.txt
pytube
opencv-python
Pillow
fpdf
